+++ src/mixins/paneMixin.ts
@@ -21,6 +21,8 @@ export default class PaneMixin extends Vue {
   mounted() {
     this.$el.id = this.paneId
 
+    this.refreshZoom()
+
     this.$nextTick(() => {
       const width = this.$el.clientWidth